@@ -5,27 +5,32 @@
   "scripts": {
     "precommit": "pretty-quick --staged && lint-staged",
     "lint": "eslint .",
-    "start": "env-cmd .env react-scripts start",
-    "start:staging": "env-cmd .env.staging react-scripts start",
+    "start": "env-cmd -f .env react-scripts start",
+    "start:staging": "env-cmd -f .env.staging react-scripts start",
     "prod": "serve -l 9000 --single ./build",
-    "build": "env-cmd .env react-scripts build",
-    "build:staging": "env-cmd .env.staging react-scripts build",
-    "test": "env-cmd .env react-scripts test --env=jsdom",
+    "build": "env-cmd -f .env react-scripts build",
+    "build:staging": "env-cmd -f .env.staging react-scripts build",
+    "test": "env-cmd -f .env react-scripts test --env=jsdom",
     "eject": "react-scripts eject",
     "now": "now --public --npm && now alias",
     "now-start": "serve --single ./build"
   },
   "dependencies": {
-    "@material-ui/core": "^3.0.1",
-    "@material-ui/icons": "^3.0.1",
-    "@material-ui/lab": "^3.0.0-alpha.19",
+    "@material-ui/core": "^3.9.3",
+    "@material-ui/icons": "^3.0.2",
+    "@material-ui/lab": "^3.0.0-alpha.30",
     "@reach/router": "^1.2.1",
     "@rematch/core": "^1.0.0-beta.4",
     "@rematch/persist": "^1.0.2",
     "apollo-boost": "^0.1.16",
     "chart.js": "^2.7.2",
     "classnames": "^2.2.6",
-    "env-cmd": "^8.0.2",
+    "env-cmd": "^9.0.3",
+    "eos-transit": "^3.1.2",
+    "eos-transit-lynx-provider": "^3.1.2",
+    "eos-transit-meetone-provider": "^3.1.2",
+    "eos-transit-scatter-provider": "^3.2.2",
+    "eos-transit-tokenpocket-provider": "^3.1.4",
     "filter-objects": "^2.1.1",
     "graphql": "^14.0.2",
     "i18next": "^12.0.0",
@@ -35,25 +40,22 @@
     "lodash.isempty": "^4.4.0",
     "lodash.set": "^4.3.2",
     "lodash.uniq": "^4.5.0",
-    "moment": "^2.22.2",
-    "qs": "^6.5.2",
+    "qs": "^6.7.0",
     "random-color-rgb": "^1.1.1",
-    "rc-slider": "^8.6.3",
-    "react": "^16.4.2",
-    "react-apollo": "^2.1.11",
-    "react-autosuggest": "^9.4.2",
-    "react-chartjs-2": "^2.7.4",
-    "react-dom": "^16.4.2",
+    "rc-slider": "^8.6.13",
+    "react": "^16.8.6",
+    "react-apollo": "^2.5.8",
+    "react-autosuggest": "^9.4.3",
+    "react-chartjs-2": "^2.7.6",
+    "react-dom": "16.8.6",
     "react-highlighter": "^0.4.3",
     "react-i18next": "^8.3.0",
     "react-placeholder": "^3.0.1",
-    "react-redux": "^5.0.7",
-    "redux": "^4.0.0",
+    "react-redux": "^7.1.0",
+    "redux": "^4.0.1",
     "redux-persist": "^5.10.0",
-    "scatterjs-core": "^2.3.2",
-    "scatterjs-plugin-eosjs2": "^1.1.4",
-    "serve": "^10.0.0",
-    "yup": "0.26.6"
+    "serve": "^11.0.2",
+    "yup": "0.27.0"
   },
   "devDependencies": {
     "eslint-config-standard": "^12.0.0",
@@ -64,7 +66,7 @@
     "lint-staged": "^7.2.2",
     "prettier": "^1.14.2",
     "pretty-quick": "^1.6.0",
-    "react-scripts": "2.0.5",
+    "react-scripts": "3.0.1",
     "snazzy": "^8.0.0",
     "standard": "^12.0.1"
   },
