@@ -3,17 +3,20 @@ import { Router } from '@reach/router'
 
 import Layout from 'components/layout'
 import NotFound from 'routes/not-found'
+import { WalletProvider } from 'hooks/wallet'
 import routes from 'routes'
 
 const App = () => (
-  <Layout>
-    <Router>
-      {routes.map(({ path, Component }) => (
-        <Component key={`path-${path}`} path={path} />
-      ))}
-      <NotFound default />
-    </Router>
-  </Layout>
+  <WalletProvider>
+    <Layout>
+      <Router>
+        {routes.map(({ path, Component }) => (
+          <Component key={`path-${path}`} path={path} />
+        ))}
+        <NotFound default />
+      </Router>
+    </Layout>
+  </WalletProvider>
 )
 
 export default App
